pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mochi-gallery"
version = "0.1.0"
description = "A Mochimo Blockchain Haiku Visualization Tool"
authors = [
  { name="Your Name", email="your@email.com" },
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai",
    "requests",
    "pillow",
    "pydantic",
    "python-dotenv"
]

[project.scripts]
# This allows you to run 'mochi-gallery' from anywhere in the terminal
mochi-gallery = "mochi_gallery.cli:main"
